move everything to example file
